refactor: change the final fixes in the entire project
