+++ src/components/PopupWithForm.js
@@ -5,12 +5,20 @@ export default class PopupWithForm extends Popup {
     super(popupSelector);
     this._submitForm = submitForm;
     this._form = this._popup.querySelector('.popup__form');
-    this.submitButton = this._form.querySelector('.popup__submit-button');
+    this._submitButton = this._form.querySelector('.popup__submit-button');
     this._inputs = Array.from(this._popup.querySelectorAll('.popup__input'));
 
     this._getInputsValues = this._getInputsValues.bind(this);
     this._submitHandler = this._submitHandler.bind(this);
     this.setInputsValues = this.setInputsValues.bind(this);
+  }
+
+  get form() {
+    return this._form;
+  }
+
+  get submitButton() {
+    return this._submitButton;
   }
 
   _getInputsValues() {
