--- src/components/utils.js
@@ -1,3 +1,0 @@
-export function renderLoading(submitButton, loadingMessage) {
-  submitButton.textContent = loadingMessage;
-}
