--- src/index.js
@@ -1,4 +1,0 @@
-import './pages/index.css';
-import main from './components/script.js';
-
-main();
