+++ webpack.config.js
@@ -4,7 +4,7 @@ const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports = {
-  entry: { main: './src/index.js' },
+  entry: { main: './src/pages/index.js' },
   output: {
     path: path.resolve(__dirname, 'dist'),
     filename: 'main.js',
@@ -19,6 +19,8 @@ module.exports = {
     port: 8080,
     open: true
   },
+
+  devtool: "source-map",
 
   module: {
     rules: [{
